domparsing/serializing_xhtml_doctypevariation2.xhtml

<!DOCTYPE  foo  SYSTEM "external.dtd"  >
<html xmlns="http://www.w3.org/1999/xhtml">
 <head>
  <title>DOM Serialization in XHTML - variation 2</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script><![CDATA[
  test(function(){
    assert_equals(new XMLSerializer().serializeToString(document.doctype), 
                  "<!DOCTYPE foo SYSTEM \"external.dtd\">");
  }, "Parsed DOCTYPE serialized correctly");
  // ]]>
  </script>
 </body>
</html>
